% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeFuelHeat.R
\name{analyzeFuelHeat}
\alias{analyzeFuelHeat}
\title{analyzeFuelHeat}
\usage{
analyzeFuelHeat(plant_list, boiler_fuel_data, dom_pct, major_pct, fuel_codes)
}
\arguments{
\item{plant_list}{vector of all plant codes from 2015, input file 2015_Plants.csv}

\item{boiler_fuel_data}{data.frame EIA-923 Monthly Boiler Fuel Consumption and Emissions Time
Series File, 2015 Final Revision, input file 2015_BoilerFuelData.csv}

\item{dom_pct}{numeric threshold for dominant fuels}

\item{major_pct}{numeric threshold for boiler majority fuel}

\item{fuel_codes}{Fuel generalization list defined as
\code{list(biomass = c("AB",'BLQ','MSB','MSN','OBL','OBS','SLW','TDF','WDS'), coal = c("ANT","BIT","LIG","PC","SUB","WC"), gas = c("BFG","LFG","NG","OBG","OG","PG","SGC","SGP"), oil = c("DFO","JF","KER","RFO","WO"), other = c("OTH"))}}
}
\value{
\code{raw_fuel_heat_df_2} annual fuel-specific fuel heat, and boiler total fuel heat at each boiler,
flags dominant fuels, majority of boiler fuel, and fuels with zero consmption,
assigns dominant fuel codes
}
\description{
Calculates annual fuel-specific fuel heat, and boiler total fuel heat at each boiler,
flags dominant fuels, majority of boiler fuel, and fuels with zero consmption,
assigns dominant fuel codes.  Returns object for output files fuel_heat_table.csv and
fuel_heat_ST_only.csv. \cr \cr
Executed By: non_nuke_condenser_duty.R \cr
}
\examples{
analyzeFuelHeat(plant_list, boiler_fuel_data_df, dom_thrsh, maj_thrsh, fuel_codes)
}
