% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeNGCCCD.R
\name{analyzeNGCCCD}
\alias{analyzeNGCCCD}
\title{analyzeNGCCCD}
\usage{
analyzeNGCCCD(bogen_netgen_df, bogen_fuel_heat_df, bogen_nom_loss_df)
}
\arguments{
\item{bogen_netgen_df}{monthly net generation by bogen, output of the assignBogen() function}

\item{bogen_fuel_heat_df}{monthly fuel heat by bogen, output from assignBogen() function}

\item{bogen_nom_loss_df}{monthly nominal losses by bogen, output of the assignBogen()}
}
\value{
\code{final_ngcc_cd} data.frame containing monthly condenser duty for general_mover = NGCC by bogen
}
\description{
calculates monthly condenser duty for general_mover = NGCC by bogen \cr \cr
Executed By: non_nuke_condenser_duty.R \cr
}
\examples{
analyzeNGCCCD(data_by_bogen$bogen_netgen, data_by_bogen$bogen_fuel_heat, data_by_bogen$bogen_nom_loss)
}
