% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSteamCD.R
\name{analyzeSteamCD}
\alias{analyzeSteamCD}
\title{analyzeSteamCD}
\usage{
analyzeSteamCD(bogen_netgen_df, bogen_steam_heat_df, bogen_nom_loss_df)
}
\arguments{
\item{bogen_netgen_df}{monthly net generation by bogen, output of the assignBogen() function}

\item{bogen_steam_heat_df}{monthly steam heat by bogen, output of the assignBogen() function}

\item{bogen_nom_loss_df}{monthly nominal losses by bogen, output of the assignBogen()}
}
\value{
\code{final_st_cd} data.frame containing monthly condenser duty for stream plants by bogen
}
\description{
calculates monthly condenser duty for stream plants by bogen \cr \cr
Executed By: non_nuke_condenser_duty.R \cr
}
\examples{
analyzeSteamCD(data_by_bogen$bogen_netgen, data_by_bogen$bogen_steam_heat, data_by_bogen$bogen_nom_loss)
}
