% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignBogen.R
\name{assignBogen}
\alias{assignBogen}
\title{assignBogen}
\usage{
assignBogen(nom_loss, netgen, fuel_heat, steam_heat, boiler_key, gen_key)
}
\arguments{
\item{nom_loss}{monthly summed nominal losses by plant and boiler, output from the
analyzeNominalLosses() function}

\item{netgen}{net generation data with negatives replaced with zeros and generators
with negative annual net generation removed}

\item{fuel_heat}{annual fuel-specific fuel heat, and boiler total fuel heat at each boiler,
flags dominant fuels, majority of boiler fuel, and fuels with zero
consmption, assigns dominant fuel codes, output of the analyzeFuelHeat()
function}

\item{steam_heat}{monthly steam heat by plant and boiler, SteamHeatanalysis() function output}

\item{boiler_key}{plant_boiler to bogen associations, from sheet3_key.csv}

\item{gen_key}{plant_gen to bogen associations, from sheet4_key.csv}
}
\value{
\code{results} list containing bogen_fuel_heat, bogen_nom_loss, bogen_netgen, bogen_steam_heat
}
\description{
calculates monthly fuel heat by bogen, assigns bogen values to the variable
data frames used in the condenser duty calculation. \cr \cr
Executed By: non_nuke_condenser_duty.R \cr
}
\examples{
assignBogen(nominal_losses, cd_net_generation$pos_netgen, fuel_heat_table, final_steam_heat,
           boiler_bogen_key, gen_bogen_key)
}
