% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignDominance.R
\name{assignDominance}
\alias{assignDominance}
\title{assignDominance}
\usage{
assignDominance(fuel_code, fuel_dominance, fuel_code_list)
}
\arguments{
\item{fuel_code}{Fuel generalization code from list defined as
\code{list(biomass = c("AB",'BLQ','MSB','MSN','OBL','OBS','SLW','TDF','WDS'), coal = c("ANT","BIT","LIG","PC","SUB","WC"), gas = c("BFG","LFG","NG","OBG","OG","PG","SGC","SGP"), oil = c("DFO","JF","KER","RFO","WO"), other = c("OTH"))}}

\item{fuel_dominance}{flagged dominant fuel, majority of boiler fuel, and fuels
with zero consmption, assigns dominant fuel codes}

\item{fuel_code_list}{Fuel generalization list defined as
\code{list(biomass = c("AB",'BLQ','MSB','MSN','OBL','OBS','SLW','TDF','WDS'), coal = c("ANT","BIT","LIG","PC","SUB","WC"), gas = c("BFG","LFG","NG","OBG","OG","PG","SGC","SGP"), oil = c("DFO","JF","KER","RFO","WO"), other = c("OTH"))}}
}
\value{
\code{result} character string assigning the dominant fuel code
}
\description{
Assigns dominant fuel codes \cr \cr
Executed By: \itemize{\item analyzeFuelHeat.R
\item fuel_heat_analysis.R} \cr
}
\examples{
assignDominance(raw_fuel_heat_df_2$Reported.Fuel.Type.Code[x], 
               raw_fuel_heat_df_2$fuel_dominance[x], fuel_code_list = fuel_codes)
}
