% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calFuelDominance.R
\name{calFuelDominance}
\alias{calFuelDominance}
\title{calFuelDominance}
\usage{
calFuelDominance(fuel_pct, dom_pct, major_pct)
}
\arguments{
\item{fuel_pct}{fuel_specific_total / boiler_total calculated using values from
EIA-923 Monthly Boiler Fuel Consumption and Emissions Time Series File,
2015 Final Revision, input file 2015_BoilerFuelData.csv}

\item{dom_pct}{numeric threshold for dominant fuels}

\item{major_pct}{numeric threshold for boiler majority fuel}
}
\value{
\code{dom_fuel} vector that flags dominant fuels, majority of boiler fuel, and fuels
with zero consmption, assigns dominant fuel codes
}
\description{
calculate dominant fuel type, flag dominant fuels, majority of boiler fuel,
and fuels with zero consmption \cr \cr
Executed By: \itemize{\item analyzeFuelHeat.R
\item fuel_heat_analysis.R} \cr
}
\examples{
sapply(raw_fuel_heat_df_2$fuel_pct, calFuelDominance, dom_pct = dom_pct, major_pct = major_pct)
}
