% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calFuelHeat.R
\name{calFuelHeat}
\alias{calFuelHeat}
\title{calFuelHeat}
\usage{
calFuelHeat(quant_consumed, unit_heat)
}
\arguments{
\item{quant_consumed}{quantity heat consumed values from EIA-923 Monthly Boiler Fuel Consumption
and Emissions Time Series File, 2015 Final Revision, input file
2015_BoilerFuelData.csv}

\item{unit_heat}{heat content for fuel (MMbtu/Unit) values from data.frame EIA-923 Monthly Boiler
Fuel Consumption and Emissions Time Series File, 2015 Final Revision, input file
2015_BoilerFuelData.csv}
}
\value{
\code{fuel_heat} annual fuel-specific fuel heat, and boiler total fuel heat at each boiler
}
\description{
calculate fuel heat by quantity consumed*heat content for fuel (MMbtu/Unit) \cr \cr
Executed By: \itemize{\item analyzeFuelHeat.R
\item fuel_heat_analysis.R} \cr
}
\examples{
as.data.frame(mapply(calFuelHeat, boiler_fuel_data[(16:27)], boiler_fuel_data[28:39]), stringsAsFactors = F)
}
