% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calNGCCCondenserDuty.R
\name{calNGCCCondenserDuty}
\alias{calNGCCCondenserDuty}
\title{calNGCCCondenserDuty}
\usage{
calNGCCCondenserDuty(fuel_heat, netgen, nominal_loss)
}
\arguments{
\item{fuel_heat}{monthly fuel heat by bogen, output from assignBogen() function}

\item{netgen}{monthly net generation by bogen, output of the assignBogen() function}

\item{nominal_loss}{monthly nominal losses by bogen, output of the assignBogen()}
}
\value{
\code{condenser_duty} data.frame containing monthly condenser duty for
general_mover = NGCC by bogen
}
\description{
calculate ngcc condenser duty \cr \cr
Executed By: analyzeNGCCCD.R \cr
}
\examples{
calNGCCCondenserDuty(cd_cal_table_2[4:15], cd_cal_table_2[28:39], cd_cal_table_2[16:27]) 
}
