% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calNukeCd.R
\name{calNukeCd}
\alias{calNukeCd}
\title{calNukeCd}
\usage{
calNukeCd(mo_fuel_heat, net_gen, percent_losses = 0.02)
}
\arguments{
\item{mo_fuel_heat}{data.frame from EIA-923 Monthly Generation and Fuel Consumption Time Series
File, 2015 Final Revision containing monthly fuel heat}

\item{net_gen}{data.frame from EIA-923 Monthly Generation and Fuel Consumption Time Series
File, 2015 Final Revision containing monthly net generation}

\item{percent_losses}{numeric value setting percent heat loss, default value 0.02}
}
\value{
\code{condenser_duty} data.frame of monthly condenser duty values for nuclear plants by plant and
nuclear unit id
}
\description{
calculate condenser duty for nuclear plants \cr \cr
Executed By: \itemize{\item NukeCDcal.R
\item NuclearCD_funs.R} \cr
}
\examples{
calNukeCd(nuke_cd[5:16], nuke_cd[17:28])
}
