% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calSteamHeat.R
\name{calSteamHeat}
\alias{calSteamHeat}
\title{calSteamHeat}
\usage{
calSteamHeat(fuel_heat, boiler_efficiency)
}
\arguments{
\item{fuel_heat}{numeric fuel heat subset of fuel_heat_table.csv, with only
Reported.Prime.Mover=="ST"}

\item{boiler_efficiency}{numeric boiler efficiency from 2015 Form EIA-860 Data - Schedule 6C,
'Boiler Information - Design Parameters', input file 2015_BoilerDesignInfo.csv}
}
\value{
\code{steam_heat} numeric steam heat from fuel heat and boiler efficiency
}
\description{
calculates monthly steam heat from fuel heat and boiler efficiency \cr \cr
Executed By: \itemize{\item estEfficiency.R
\item steamHeatAnalysis.R} \cr
}
\examples{
as.data.frame(apply(steam_heat_rpt_eff[5:16], 2, calSteamHeat, boiler_efficiency = steam_heat_rpt_eff$boiler_efficiency))
}
