% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd-package.R
\docType{package}
\name{cd-package}
\alias{cd}
\alias{cd-package}
\title{cd: Climate Departure Analysis from ERA5-Land Reanalysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Compute climate anomalies from ERA5-Land reanalysis data for custom areas of interest. Producer functions download ERA5-Land monthly means, derive variables (VPD, RH, soil moisture), and publish Cloud-Optimized GeoTIFFs with a static STAC catalog. Consumer functions query the catalog, crop to AOI, compute baselines and anomalies for arbitrary reference periods, and run trend analysis (Mann-Kendall and Theil-Sen).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NewGraphEnvironment/cd}
  \item \url{https://newgraphenvironment.github.io/cd/}
  \item Report bugs at \url{https://github.com/NewGraphEnvironment/cd/issues}
}

}
\author{
\strong{Maintainer}: Allan Irvine \email{al@newgraphenvironment.com} (\href{https://orcid.org/0000-0002-3495-2128}{ORCID})

Other contributors:
\itemize{
  \item New Graph Environment Ltd. [copyright holder]
}

}
\keyword{internal}
