% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_cache.R
\name{cd_cache_clear}
\alias{cd_cache_clear}
\title{Clear the cd cache}
\usage{
cd_cache_clear(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character. Override the default cache location.}
}
\value{
Invisibly returns the number of files removed.
}
\description{
Removes all cached files from the cd cache directory.
}
\examples{
\dontrun{
cd_cache_clear()
}

}
