% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_cache.R
\name{cd_cache_path}
\alias{cd_cache_path}
\title{Get cd cache directory path}
\usage{
cd_cache_path(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Character. Override the default cache location.
If NULL, uses \code{rappdirs::user_cache_dir("cd")}.}
}
\value{
Character path to the cache directory.
}
\description{
Returns the path to the cd cache directory. Creates it if
it doesn't exist.
}
\examples{
cd_cache_path()

}
