% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_catalog.R
\name{cd_catalog_default}
\alias{cd_catalog_default}
\title{Default STAC catalog URL}
\usage{
cd_catalog_default()
}
\value{
Character URL.
}
\description{
Returns the default S3-hosted STAC catalog URL for the cd package.
Override with \code{options(cd.catalog_url = "...")}.
}
\examples{
cd_catalog_default()

}
