% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_periods.R
\name{cd_periods}
\alias{cd_periods}
\title{Valid temporal aggregation periods}
\usage{
cd_periods(include_monthly = FALSE)
}
\arguments{
\item{include_monthly}{Logical. If \code{TRUE}, appends the 12 calendar
months (Jan, Feb, ..., Dec) to the seasonal periods. Default \code{FALSE}.}
}
\value{
Character vector of period names.
}
\description{
Returns the set of temporal periods supported by the cd package.
Used for input validation and iteration in extraction and analysis
functions.
}
\examples{
cd_periods()
cd_periods(include_monthly = TRUE)

}
