% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_aggregate.R
\name{cd_seasons}
\alias{cd_seasons}
\title{Default meteorological season definitions}
\usage{
cd_seasons()
}
\value{
Named list of integer vectors.
}
\description{
Returns a named list of month numbers for standard meteorological
seasons. Override to define custom seasons (e.g., wet/dry, growing
season).
}
\examples{
cd_seasons()

# Custom: hydrological year
list(wet = c(10, 11, 12, 1, 2, 3), dry = 4:9)

}
