% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dependencies.R
\name{check_dependencies}
\alias{check_dependencies}
\title{check_dependencies}
\usage{
check_dependencies()
}
\description{
checks if required packages are installed, and askes user if they want to
install and installs if Y \cr \cr
Executed By: ThermoElectricWaterConsumptionModels.R \cr
}
\examples{
check_dependencies()
}
