% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estEfficiency.R
\name{estEfficiency}
\alias{estEfficiency}
\title{estEfficiency}
\usage{
estEfficiency(fuel_heat_df, boiler_design_data_df)
}
\arguments{
\item{fuel_heat_df}{data.frame subset of fuel_heat_table.csv, with only
Reported.Prime.Mover=="ST"}

\item{boiler_design_data_df}{data.frame 2015 Form EIA-860 Data - Schedule 6C,
'Boiler Information - Design Parameters', input file 2015_BoilerDesignInfo.csv}
}
\value{
\code{results} list of estimated boiler efficiencies and reported steam heat efficiencies
}
\description{
get reported boiler efficiencies and estimate boiler efficiencies \cr \cr
Executed By: steamHeatAnalysis.R \cr
}
\examples{
estEfficiency(fuel_heat_df, boiler_design_data_df)
}
