% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcEnvironVars.R, R/getEnvVars.R
\name{getEnvVars}
\alias{getEnvVars}
\title{getEnvVars}
\usage{
getEnvVars(
  plant_list,
  dbwb_data_df,
  ws_data_df,
  wt_data_df,
  outputData_path,
  getEnvVars = T
)

getEnvVars(
  plant_list,
  dbwb_data_df,
  ws_data_df,
  wt_data_df,
  outputData_path,
  getEnvVars = T
)
}
\arguments{
\item{plant_list}{vector of all plant codes from 2015, input file 2015_Plants.csv}

\item{dbwb_data_df}{NOAA dry-bulb and wet-bulb air temperature data for the nearest three weather
stations to each power plant.  Used to calculate a distance-weighted average
of the three nearest weather stations to determine DB and WB monthly air temps
for each plant, from the 3sta_DBWB.csv input file}

\item{ws_data_df}{NOAA wind speed data for the nearest three weather stations to each power plant.
Used to calculate a distance-weighted average of the three nearest weather stations
to determine monthly wind speeds for each plant, from 3sta_WS.csv input file}

\item{wt_data_df}{Monthly kriged-estimated water temps for each plant, from the 2015_WT_Predictions.csv
input file}

\item{outputData_path}{character string indicating path for output files}

\item{getEnvVars}{TRUE/FALSE whether or not to execute getEnvVars() function, default value TRUE}
}
\value{
\code{result} data.frame that combines and format monthly environmental variables used by the TOWER
and FEWSR models (dry bulb, wet bulb, wind speed and natural water temperature)

\code{result} data.frame that combines and format monthly environmental variables used by the TOWER
and FEWSR models (dry bulb, wet bulb, wind speed and natural water temperature)
}
\description{
combines and format monthly environmental variables used by the TOWER and FEWSR models
(dry bulb, wet bulb, wind speed and natural water temperature) \\cr \\cr
Outputs the following files: environmental_variables.csv - monthly environmental variables used by the
TOWER and FEWSR models
Executed By: CondenserDutyModel_new.R \\cr

combines and format monthly environmental variables used by the TOWER and FEWSR models
(dry bulb, wet bulb, wind speed and natural water temperature) \cr \cr
Outputs the following files: environmental_variables.csv - monthly environmental variables used by the
TOWER and FEWSR models
Executed By: CondenserDutyModel_new.R \cr
}
\examples{
getEnvVars(modeled_plants$Plant.Code, dbwb_data, ws_data, wt_data, getEnvVars=T)
getEnvVars(modeled_plants$Plant.Code, dbwb_data, ws_data, wt_data, getEnvVars=T)
}
