% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHere.R
\name{getHere}
\alias{getHere}
\title{getHere}
\usage{
getHere()
}
\description{
function to initiate a .here location for the thermoelectric-water-consumption-model
scripts if no thermoelectric project is open.  If another Rproject is open, and error
message will generated telling the user to switch or close the open project \cr \cr
Executed By: ThermoElectricWaterConsumptionModels.R \cr
}
\examples{
getHere()
}
