% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_boiler_design_data.R
\name{import_boiler_design_data}
\alias{import_boiler_design_data}
\title{import_boiler_design_data}
\usage{
import_boiler_design_data(boiler_design_data_file_path)
}
\arguments{
\item{boiler_design_data_file_path}{path to ~/thermoelectric-water-consumption-models/
CondenserDutyModel/Data/2015_BoilerDesignInfo.csv}
}
\value{
\code{boiler_data} formatted 2015 Form EIA-860 Data - Schedule 6C,
'Boiler Information - Design Parameters' data
}
\description{
import and format for use the
~/thermoelectric-water-consumption-models/CondenserDutyModel/Data/2015_BoilerDesignInfo.csv
file \cr \cr
Executed By: CondenserDutyModel_new.R \cr
}
\examples{
import_boiler_design_data(paste0(inputData_path,'/2015_BoilerDesignInfo.csv'))
}
