% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_generator_data.R
\name{import_generator_data}
\alias{import_generator_data}
\title{import_generator_data}
\usage{
import_generator_data(generator_data_file_path)
}
\arguments{
\item{generator_data_file_path}{path to ~/thermoelectric-water-consumption-models/
CondenserDutyModel/Data/2015_GeneratorData.csv}
}
\value{
\code{generator_data} formatted plant generator data
}
\description{
import and format for use the
~/thermoelectric-water-consumption-models/CondenserDutyModel/Data/2015_GeneratorData.csv
file \cr \cr
Executed By: CondenserDutyModel_new.R \cr
}
\examples{
import_generator_data(paste0(inputData_path,'/2015_GeneratorData.csv'))
}
