% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sheet1_gen_and_fuel_data.R
\name{import_sheet1_gen_and_fuel_data}
\alias{import_sheet1_gen_and_fuel_data}
\title{import_sheet1_gen_and_fuel_data}
\usage{
import_sheet1_gen_and_fuel_data(plant_Data_file_path)
}
\arguments{
\item{plant_Data_file_path}{path to ~/thermoelectric-water-consumption-models/
CondenserDutyModel/Data/2015_GenerationAndFuel.csv}
}
\value{
\code{sheet1_gen_fuel_data} formatted plant generator fuel data
}
\description{
import and format for use the
~/thermoelectric-water-consumption-models/CondenserDutyModel/Data/2015_GenerationAndFuel.csv
file \cr \cr
Executed By: CondenserDutyModel_new.R \cr
}
\examples{
import_sheet1_gen_and_fuel_data(paste0(inputData_path,'/2015_GenerationAndFuel.csv'))
}
