% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeCD_ENV.R
\name{mergeCD_ENV}
\alias{mergeCD_ENV}
\title{mergeCD_ENV}
\usage{
mergeCD_ENV(combined_duties_df, environmental_variables_df, outputData_path)
}
\arguments{
\item{combined_duties_df}{monthly condender duty for all plant types by plant and cooling type}

\item{environmental_variables_df}{monthly environmental variables used by the TOWER and FEWSR models}

\item{outputData_path}{character string indicating path for output files}
}
\value{
\code{duties_envVar_final} data.frame that combines condenser duty output and environmental data
and formatted for input into the TOWER and FEWSR models
}
\description{
combines condenser duty output and environmental data and formatted for input into
the TOWER and FEWSR models \cr \cr
Outputs the following files: rawInput_FEWSR_TOWER.csv - condenser duty output and environmental data
and formatted for input into the TOWER and FEWSR models
Executed By: CondenserDutyModel_new.R \cr
}
\examples{
mergeCD_ENV(no_dupe_final_CD_w_nukes_cooling, environmental.variables)
}
