% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_nuke_condenser_duty.R
\name{non_nuke_condenser_duty}
\alias{non_nuke_condenser_duty}
\title{non_nuke_condenser_duty}
\usage{
non_nuke_condenser_duty(
  plant_list,
  boiler_fuel_data_df,
  boiler_design_data_df,
  sheet1_gen_fuel_data_df,
  sheet4_gen_fuel_data_df,
  dom_thrsh,
  maj_thrsh,
  fuel_codes,
  pub_efficency_table,
  use_published_efficiencies = T,
  boiler_bogen_key,
  gen_bogen_key,
  outputData_path
)
}
\arguments{
\item{plant_list}{vector of all plant codes from 2015, input file 2015_Plants.csv}

\item{boiler_fuel_data_df}{data.frame EIA-923 Monthly Boiler Fuel Consumption and Emissions Time
Series File, 2015 Final Revision, input file 2015_BoilerFuelData.csv}

\item{boiler_design_data_df}{data.frame 2015 Form EIA-860 Data - Schedule 6C,
'Boiler Information - Design Parameters', input file 2015_BoilerDesignInfo.csv}

\item{sheet1_gen_fuel_data_df}{data.frame EIA-923 Monthly Generation and Fuel Consumption Time
Series File, 2015 Final Revision, input file 2015_GenerationAndFuel.csv}

\item{sheet4_gen_fuel_data_df}{data.frame EIA-923 Monthly Generating Unit Net Generation Time
Series File, 2015 Final Revision, input file 2015_GenerationData.csv}

\item{dom_thrsh}{numeric threshold for dominant fuels}

\item{maj_thrsh}{numeric threshold for boiler majority fuel}

\item{fuel_codes}{Fuel generalization list defined as
\code{list(biomass = c("AB",'BLQ','MSB','MSN','OBL','OBS','SLW','TDF','WDS'), coal = c("ANT","BIT","LIG","PC","SUB","WC"), gas = c("BFG","LFG","NG","OBG","OG","PG","SGC","SGP"), oil = c("DFO","JF","KER","RFO","WO"), other = c("OTH"))}}

\item{pub_efficency_table}{list of published weighted boiler efficiencies defined as
pub_efficiency <- list("0.781596394425121" = "biomass",
"0.873790282218364" = "coal",
"0.861342667396523" = "gas",
"0.886779729944024" = "oil")}

\item{use_published_efficiencies}{TRUE/FALSE indicating whether or not to use published efficiencies
default value TRUE}

\item{boiler_bogen_key}{data.frame plant_boiler to bogen associations, input file sheet3_key.csv}

\item{gen_bogen_key}{data.frame plant_gen to bogen associations, input file sheet4_key.csv}

\item{outputData_path}{character string indicating path for output files}
}
\value{
\code{cd_with_exports} data.frame of CondenserDutyModel results by bogen for non-nuclear plants,
saved as output file ~/thermoelectric-water-consumption-models/CondenserDutyModel/Output/net_condenser_duty.csv
}
\description{
condenser duty model execution for non nuclear plants. \cr \cr
Outputs the following files:\itemize{\item fuel_heat_table.csv - annual fuel-specific fuel heat, and boiler total fuel heat at each boiler
\item fuel_heat_ST_only.csv - subset of fuel_heat_table.csv, with only Reported.Prime.Mover=="ST"
\item plant_exports.csv - monthly heat exports by plant
\item final_steam_heat.csv - monthly steam heat by plant and boiler
\item pos_netgen.csv - annual and montly process net generation by plant and generator
\item nominal_losses.csv - monthly summed nominal losses by plant and boiler
\item steam_bogen_cd.csv - monthly condenser duty for stream plants by bogen
\item ngcc_bogen_cd.csv - monthly condenser duty for general_mover = NGCC by bogen
\item bogen_fuel_heat.csv - monthly fuel heat by bogen
\item bogen_nominal_losses.csv - monthly nominal losses by bogen
\item bogen_netgen.csv - monthly net generation by bogen
\item bogen_steam_heat.csv - monthly steam heat by bogen
\item net_condenser_duty.csv - monthly net condenser duty by bogen non-nuclear plants
\item total_condenser_duty.csv - monthly total condenser duty by bogen non-nuclear plants} \cr
Executed By: CondenserDutyModel_new.R \cr
}
\examples{
non_nuke_CD_summary <- non_nuke_condenser_duty(modeled_plants$Plant.Code, boiler_fuel_data, boiler_design_data, sheet1_gen_fuel_data, 
                                              sheet4_gen_fuel_data, 0.8, 0.5, fuel.codes, pub_efficiency,
                                              use_published_efficiencies = T, page3_bogen_key, page4_bogen_key)
}
