% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processExports.R
\name{processExports}
\alias{processExports}
\title{processExports}
\usage{
processExports(plant_list, plant_exports_df, non_nuke_cd_df)
}
\arguments{
\item{plant_list}{vector of all plant codes from 2015, input file 2015_Plants.csv}

\item{plant_exports_df}{monthly heat exports by plant, output of the calHeatExports() function
in the heatExport_fun.R}

\item{non_nuke_cd_df}{combined steam and NGCC condenser duty output of analyzeSteamCD() and analyzeNGCCCD()}
}
\value{
\code{non_nuke_cd_exports} data.frame containing monthly net condenser duty heat exports for
non-nuclear plants, output of the processExports() function
}
\description{
calculates monthly net condenser duty heat exports for non-nuclear plants \cr \cr
Executed By: non_nuke_condenser_duty.R \cr
}
\examples{
processExports(plant_list, plant_exports, non_nuke_cd)
}
