% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steamHeatAnalysis.R
\name{SteamHeatanalysis}
\alias{SteamHeatanalysis}
\title{SteamHeatanalysis}
\usage{
SteamHeatanalysis(
  fuel_heat_df,
  boiler_design_data_df,
  published_eff,
  use_published_efficiencies = T
)
}
\arguments{
\item{fuel_heat_df}{data.frame subset of fuel_heat_table.csv, with only
Reported.Prime.Mover=="ST"}

\item{boiler_design_data_df}{data.frame 2015 Form EIA-860 Data - Schedule 6C,
'Boiler Information - Design Parameters', input file 2015_BoilerDesignInfo.csv}

\item{published_eff}{list of published weighted boiler efficiencies defined as
list("0.781596394425121" = "biomass",
"0.873790282218364" = "coal",
"0.861342667396523" = "gas",
"0.886779729944024" = "oil")}

\item{use_published_efficiencies}{TRUE/FALSE indicating whether or not to use published efficiencies
default value TRUE}
}
\value{
\code{steam_heat_results} data.frame monthly steam heat by plant and boiler
}
\description{
calculates monthly steam heat by plant and boiler, for output file
final_steam_heat.csv \cr \cr
Executed By: non_nuke_condenser_duty.R \cr
}
\examples{
SteamHeatanalysis(fuel_heat_ST_only, boiler_design_data, pub_efficency_table, use_published_efficiencies)
}
